Refactor layout. Should be close to working code.
